# packages/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bijou AI WhatsApp Enterprise",
    "version": "2.2.0",
    "description": "Enterprise-grade AI customer support for WhatsApp with multi-language support, human escalation, and Google Sheets integration. Built for property agents and SMEs.",
    "contact": {
      "name": "Bijou AI Support Team",
      "email": "support@mybijou.xyz",
      "url": "https://app.mybijou.xyz/static/help.html"
    },
    "license": {
      "name": "Commercial License",
      "url": "https://mybijou.xyz/terms"
    }
  },
  "paths": {
...
    "/api/knowledge/upload": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Upload Knowledge Document",
        "description": "Upload a knowledge document for a tenant.\n\nSupported formats: PDF, DOCX, TXT, MD, CSV, JSON\n\nHeaders:\n    X-Tenant-ID: Tenant UUID\n\nReturns:\n    UploadResponse with document details",
        "operationId": "upload_knowledge_document_api_knowledge_upload_post",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_knowledge_document_api_knowledge_upload_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/list": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "List Knowledge Documents",
        "description": "List all knowledge documents for a tenant.\n\nHeaders:\n    X-Tenant-ID: Tenant UUID\n\nReturns:\n    DocumentListResponse with list of documents",
        "operationId": "list_knowledge_documents_api_knowledge_list_get",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/{document_id}": {
      "delete": {
        "tags": [
          "knowledge"
        ],
        "summary": "Delete Knowledge Document",
        "description": "Delete a knowledge document.\n\nPath Parameters:\n    document_id: Document UUID\n\nHeaders:\n    X-Tenant-ID: Tenant UUID\n\nReturns:\n    DeleteResponse with success status",
        "operationId": "delete_knowledge_document_api_knowledge__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/knowledge/combined": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Get Combined Knowledge",
        "description": "Get combined knowledge text from all documents for a tenant.\n\nThis is used to inject knowledge into the AI's context.\n\nHeaders:\n    X-Tenant-ID: Tenant UUID\n\nReturns:\n    CombinedKnowledgeResponse with combined text",
        "operationId": "get_combined_knowledge_api_knowledge_combined_get",
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CombinedKnowledgeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_knowledge_document_api_knowledge_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_knowledge_document_api_knowledge_upload_post"
      },
...
      "CombinedKnowledgeResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "combined_text": {
            "type": "string",
            "title": "Combined Text"
          },
          "document_count": {
            "type": "integer",
            "title": "Document Count"
          },
          "total_length": {
            "type": "integer",
            "title": "Total Length"
          }
        },
        "type": "object",
        "required": [
          "success",
          "tenant_id",
          "combined_text",
          "document_count",
          "total_length"
        ],
        "title": "CombinedKnowledgeResponse",
        "description": "Combined knowledge text from all documents"
      },
...
      "DeleteResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "document_id",
          "message"
        ],
        "title": "DeleteResponse",
        "description": "Response after document deletion"
      },
      "DocumentInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "content_length": {
            "type": "integer",
            "title": "Content Length"
          },
          "uploaded_at": {
            "type": "string",
            "title": "Uploaded At"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "file_type",
          "content_length",
          "uploaded_at",
          "metadata"
        ],
        "title": "DocumentInfo",
        "description": "Information about a knowledge document"
      },
      "DocumentListResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentInfo"
            },
            "type": "array",
            "title": "Documents"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          }
        },
        "type": "object",
        "required": [
          "success",
          "tenant_id",
          "documents",
          "total_count"
        ],
        "title": "DocumentListResponse",
        "description": "List of documents for a tenant"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "content_length": {
            "type": "integer",
            "title": "Content Length"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success",
          "document_id",
          "filename",
          "file_type",
          "content_length",
          "message"
        ],
        "title": "UploadResponse",
        "description": "Response after document upload"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}